FIX: moved order list to render method
Removed assigning order list to variable and
moved it directly to render method.

Arek's advice: If you do not change the
reference of the variable then use const.
Also extracting this orders.map to variable
makes visual noise, as part of the rendering
is "floating" above the actual return.
Consider making ordersList class method or
inlining it in line 28.